requirements.txt
# Core dependencies for Cube Analyst
autogen==0.9.9
autogen-agentchat>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit==1.49.1

# Additional dependencies that AutoGen might need
pydantic>=2.0.0
httpx>=0.25.0
typing-extensions>=4.5.0
nest_asyncio>=1.6.0
